Ensure PDF_SUBHEADING defaults to an empty string
Relates to story: https://trello.com/c/utAUt3qG/1253-add-submission-configuration-serviceoutputemail-to-editor

At present, the PDF_SUBHEADING can be left blank in the email submission settings form.
The editor will detect this and remove the PDF_SUBHEADING value from the database.
This means the submitter will throw am 'Unproccessable entity' error as pdf_subheading is nil, rendering the schema invalid.

This commit ensures the value of PDF_SUBHEADING will not be nil.