Support content pages
This updates the navigation spec to include the new content page.

Content pages have no ability to capture a user input as they are there
for informative purposes only. However they can still be part of a flow
for a form and therefore still have the Continue button, that does a
POST as expected but no user data is saved.

Therefore there is no expectation for the content page to appear in the
check your answers page nor be part of a submission that is processed by
the platform.