Redirect users to start page once session has expired
After the current 30 minutes of inactivity the session is removed. We
want the user to be redirected to the start page when they next interact
with the form.

We also allow entry into any standalone page without first having to
visit the start page. These standalone pages are not considered part of
a form flow so it is reasonable that a user may want to visit that page
independently of the form and not have to go to the start page first.

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>
Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>