Allow footer to move dynamically with page content
We previously stuck the footer to the bottom of the page. While this
looked nice it did also create an issue for any user that would zoom
into the page a little too much.

Therefore, remove the CSS which makes the footer stick and instead allow
it to move dynamically based on the content of a page. We then make the
background html element the same colour as the footer for the times when
there is not very much content on the page.

Users should now be able to zoom in and the content won't be obscured
by the footer.