Fix golang dependency installation issue
We install golang onto the container in order to have easy access to
the Vegeta package for load testing. It was failing installation during
the build process because of a `undefined: tls.Dialer` error.

One solution is for us to remove golang entirely as we only use it for
load testing.

However it seems that copying over the golang binary from the 1.16
golang alpine image directly into our container seems to solve the
installation issue.