Exclude health endpoint from sentry
/health is used by Kubernetes to check whether the container is ok. We
do not need this to make it to Sentry.

Also exclude the robots.txt endpoint.

Following the docs here:

https://docs.sentry.io/platforms/ruby/configuration/sampling/

Returning true would set the sample rate to 1 which means every request
is sent to Sentry. Returning false sets it to 0 so the the transaction
is not sent.

`sentry: [Tracing] Discarding <rails.request> transaction </health> because traces_sampler returned 0 or false`

In the future it may be necessary to adjust the sample rate so that it
is not 100% of all transactions, but for the purposes of the trial it
will do for now.