Use allowed file types from component metadata
Going forward we need to be able to take the allowed file types for a
given component from the metadata and pass that onto the filestore when
required as the filestore is the one that checks the types.

The default list of allowed file types should also be generated at the
same time as a new file upload component is created.