Attempt standard decode64 if strict_decode64 fails
In Ruby's standard library when using standard encode64 on a string CR
and LF characters are injected into it. This was causing the encrypted-user-id-and-token
header that is checked during JWT authentication between the submitter
and the filestore to fail with and ArgumentError. It is not able to
correctly decode the CR and LF characters.

In order get round this we changed the encoding mechanism to use
strict_encode64 which resolves that issue. However we need to be able to
cater for both when we deploy as there will still be users
interacting with the forms which will have encoded the encrypted-user-id-and-token
using the standard encode64 method.

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>
Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>