Fix optional file upload components
When a file upload component is optional what is actually saved in the
user_data object is a `{}`. e.g `{ 'dog-picture_upload_1' => {} }`

This is ok as when we get to the summary we will want to show that there
was a file upload question, but since it's optional there may or may not
be a file name.

However we cannot send `nil` in the submission payload as this fails
schema validation in the Submitter. So instead we send an `''`. The PDF
generator should then correctly iterate over the answer and just present
the name of the question but with no name of the file.

When it comes to the attachments we should just not include the file
object associated with that question at all if the user hasn't actually
uploaded a file.

Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>