Cater for optional checkbox answers
This fixes a for when checkboxes are optional and a user does not pick
any of the choices.

If the user had not previously answered the question and had proceeded
through the rest of the form and submitted this would be fine.

However in the event the user comes back to the checkboxes page and
unticks all of the items then this was not reflected in the saved
answers. This is because the frontend has no concept of what the user
has saved at that point and does not post anything relating to that
answer to the backend at all. Therefore their previous answer was never
overwritten.

In those instances where there is no answer for a checkbox page we set
an empty array against that components corresponding ID in their user
answers therefore overwriting any previous answer that they may have
had.