Run acceptance tests relating to this version of the runner
The fb-acceptance-tests repo holds all the tests related to the legacy
runner as well as the new v2 runner.

Running make spec-ci spins up all docker containers relating to the
legacy runner as well as running all of those tests.

For the fb-runner app there is only the new_runner_spec.rb so we may as
well only run this.

Also we don't need to use docker to run these tests.

This reduces the acceptance tests run from 5 minutes down to about 1
minute.