Send answered pages based on branching
The traversed pages object will traverse to all pages and
evaluates the branches and selects the answers from user
data **whatever** the user data contains.

The user data can contain all answers for all pages in all branches!
But it is up to the `TraversedPages` object to decide which one
will be include in the submission.