Generate user_id to replace session_id when talking with platform apps
Currently the session_id is used as the user_id:

fb-runner/encrypted_user_id_and_token.rb at f1a0c3283c9abdc36ee952e7a5d586860ac2a87a · ministryofjustice/fb-runner

That user_id (session_id) is then used for

the URL for both the datastore and the filestore.
the subject in the JWT payload.
the naming of files that get saved to the S3 bucket
used in the encrypted_user_id_and_token that gets sent to the Submitter for use when it has to
The user_token is generated when a user visits the runner for the first time:

fb-runner/user_data.rb at 348f50d7d1bddd4550a9309e0089aba3aaebc72a · ministryofjustice/fb-runner

It is probably inherently insecure that we use the session_id as part of the URL.

The legacy runner works as follows:

https://github.com/ministryofjustice/moj-forms-tech-docs/blob/main/source/documentation/services/runner.html.md.erb#L21

It generates a random user_id which is then used in the same manner across all the runners. This is effectively the same thing we do for the user_token so we should take the same approach.

Backwards compatibility should be maintained. If a user has already visited the runner and is already using the session_id as the user_id then this should be maintained for the duration of their session.