Send the user id in file url to submitter
Use the user id always (when uploading and when asking to submitter
to download) and If we don't have use the session_id.

If we don't do this when the Submitter downloads the file,
the user_id and session_id won't match. Conclusion: it should be
the same id that you sent the file.

When we send the file to filestore we attach an user id to it.
When we pass to submitter the attachment the id should be the same
one that you sent. Given that this PR fix this issue.

Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>
Co-authored-by: Brendan Butler <brendan.butler@digital.justice.gov.uk>