Enable MoJ Forms GA4 in Test
This adds the MoJ Forms GA4 measurement ID to an initialiser but only
for the Test environment currently.

The Presenter looks for the existence of this and will then inject the
correct analytics scripts into the head.

We are setting this in an initialiser instead of as an environment
variable as we want to be able to roll this change out to all currently
published runners without the need for republishing.