Add .dockerignore file
There are a lot of directories and dotfiles that we do not need to be
copied over to the app. This will be especially true once acceptance
tests are moved into this repo.

We add the vendor and node_modules folder as the Dockerfile will do the
dependency install and manually copy those over.