Add the ability to send CSV of user answers as an attachment
Setting `SERVICE_CSV_OUTPUT` to 'true' (or just having it present) in
the form container means that an additional action will be added to the
payload that is sent to the Submitter.

The Submitter will then action that and convert the user answers into a
CSV and then attach it to a separate email to the form owner.