Load autocomplete items into select list
Autocomplete items will be injected into each forms container and set on
an environment variable called AUTOCOMPLETE_ITEMS. We set this on the
configuration object for the app making it available to be loaded into
any autocomplete page by finding the correct items for a given component
by it's uuid.

For local development we can override this by passing in an environment
variable called AUTOCOMPLETE_FIXTURE. By default this is set to use the
`countries.json` fixture which is found in the MetadataPresenter gem.

Also the value that an end user chooses to submit for an autocomplete
component is sent as a JSON string represenation with both the 'text'
and 'value'. This is because for the public citizen user we want to show
the more human readable form, which is usually the 'text' property, but
when we are submitting their answers to a Form Owner we want to pass the
'value' property into the submission payload.