Convert checkbox component answers to semi colon separated string
Previously checkbox answers were just an array of the different options
when submitted. e.g

['beef, cheese, tomato', 'chicken, tomato, lettuce']

This changes the answers to take the form:

'beef, cheese, tomato; chicken, tomato, lettuce'

The answers are concatenated into a string and separated by a semi
colon.

This changes the output for both the PDF file of the answers and the CSV
attachment generated by the submitter.

Co-authored-by: Hellema Ibrahim <hellema.ibrahim@digital.justice.gov.uk>