Add ability to send a confirmation email of users answers
The confirmation email feature requires the user's answer to an
email component to work. Therefore, we need to add a new action
that pulls out the user's answer from the correct email component,
this is identified by the 'CONFIRMATION_EMAIL_COMPONENT_ID'.

The presence of a 'CONFIRMATION_EMAIL_COMPONENT_ID' will add the
action to the payload that is sent to the Submitter. The Submitter
then actions this, sending a copy of the user's answers to the email
address provided by the email component.

Since, it is possible for a form editor to choose an email component
that is optional, we need to ensure we the submitter does not attempt
to send an email if there is no email provided.