Download service metadata from S3
We now upload the service metadata and autocomplete items as objects to
an S3 bucket. Should there be a SERVICE_ID in the environment of the
form container then make the request to S3 to get the necessary files.

Then load them in much the same way as the current metadata is loaded.

We need to maintain backwards compatability to allow forms already
published to look for metadata in the environment, as well as local
development's use of fixtures