Passing reference number to submitter payload
Adding the reference number in the user data object with the key
'moj_forms_reference_number'. We have chosen to name this key like this
because we do not think a form user would create a component with the
same name. We have used the reference number in the submitter payload to
substitute the text {{reference_number}} injected by the editor