Add reference number to submitter payload
There is a need for the reference number for a submission to be injected
into other types of payloads when being processed by the submitter.
Until now the reference number has only been present in some of the
payload strings such as email subjects or email bodies.

Set the reference number as an attribute on the meta payload, should it
be present in the user_data, and then send that along with the other
meta attributes to the submitter.

We do not add it as part of the user answers as technically the user did
not add it themselves, and also there are no component id or component
field attributes relating to the reference number since it does not have
its' own component.