Confirmation email bug
We have a bug whereby if the form editor uses the same email for
the `ENV['SERVICE_EMAIL_OUTPUT']` and the
`user_data[ENV['CONFIRMATION_EMAIL_COMPONENT_ID']]`, only the
service email will be sent.

This is because, when the submitter app receives the submitter
payload form the Runner, it only sends emails that are unique
based on the 'to' email address and the 'attachments'.
Since in the case of the service email and the confirmation email,
both 'to' and 'attachments' will be identical, only one email will
be sent.

This bug therefore, only affects form editors that would use the
same 'to' email for testing purposes.

To mitigate this bug, we can append a '+confirmation' string to the
email address that is submitted for the confirmation email.