Reply To - Prepend service name to from address
Currently, when an end user receives a confirmation email from
a service, the from email address is displayed or (if the service
is using the default from address) the default moj forms email
address is displayed.

We would like to prepend the email address in the 'from' and the
'reply-to' with the service name. This is so the emails received
appear as though they are coming from the service.