Reply-To: Allow backwards compatibility for reply-to/from address
We have changed the functionality of the from address in the Editor
to now inject a 'SERVICE_EMAIL_ REPLY_TO' env var. To ensure backwards
compatibility we need to check whether a 'REPLY_TO' variable exists,
if it does we should use it. If it does not, we can use the
'SERVICE_EMAIL_FROM'.