Prepend service name to from address and reply-to addresses
Currently, when an end user receives a confirmation email from
a service, the email address is displayed.

We would like to prepend the email address in the default 'from'
and the 'reply-to' with the service name. This is so the emails
received appear as though they are coming from the service.