Add service_slug_config method
We now require this method as it is called in the metadata
presenter gem.

The service slug used is in our submission mechanism and the user
datastore adapter. We also require the service slug for our save
and return feature.

The SERVICE_SLUG is injected into the runner container from the
Editor.