Deleting a non-existent file will gracefully re-render
In the case when a file causes an error, eg: wrong file type,
the file is visually added to the page but is not present in the
datastore. When a user attempts to delete the file, the app
breaks as it cannot find the file.

To prevent this, we just want to add an early return.