Add additional context to date answer exceptions (#1299)
* Add additional context to date answer exceptions

On the back of this sentry exception:
https://ministryofjustice.sentry.io/issues/4612935516

Added some additional context that will be attached to the sentry exception so we can investigate difficult to diagnose exceptions occurring with the parsing of some answers, in this case date answers.

The code can handle exceptions in other component types, those defined as methods, so: date, checkboxes, upload, multiupload and autocomplete. But realistically only date and perhaps autocomplete may raise exceptions, as the other components are trivial.

More info:
https://docs.sentry.io/platforms/ruby/enriching-events/context/

* Use a string instead of answer object

Sentry might not be able to parse the object, so it should be safer to just use a string (using inspect).