Customise timeouts and propagate request id (#1304)
* Customise timeouts and propagate request id

Faraday can be customised with different timeouts for open and read operations. Open operation can have a small timeout, set to 10 seconds but if we observe any issues we can tweak this.

Read timeout can be configured per adapter. I've left all adapters to 30 secs (as before) except the filestore adapter that I've increased to 45 secs in a attempt to reduce timeouts errors we get from time to time. However this may still be not enough so we can observe and tweak if necessary.

Finally, probably the most important thing, we are now propagating the request Id, so each service (filestore, datastore, etc.) would get it and we can correlate requests from the runner to each of these services, which is very useful when looking at the logs.

* Fix tests