Add debug to help diagnose unfiltered params issue
We are having some random exceptions:

```ActionController::UnfilteredParameters
unable to convert unpermitted parameters to hash```

I can't find a clear source of the issue, other than the obvious stated by the exception that the params have not been filtered, however the code performs filtering.

Adding a bit of debug context to the sentry alert to try to diagnose the issue.