Workaround for random unfiltered params error
We couldn't find the source of the problem, why for some reason a few requests a day produce unfiltered params error.

So far all I can see is we are filtering (in the metadata presenter) the params, and permitting the answers (`params[:answers].permit!`)

Why somehow in some scenarios this permitted params object become unpermitted or contains unpermitted params is still a mistery.

This workaround does not fix whatever is the underlying issue but fix the result of it by re-permitting the answer params, which in the debug introduced previously it always seems to be 1 upload param.